+++ Frontend/src/components/header/header.consts.js
@@ -1,0 +1,1 @@
+export const STATICPLACEHOLDER = 'Search for services';
