+++ Frontend/src/components/serviceDropdown/service-dropdown.consts.js
@@ -1,0 +1,1 @@
+export const TRENDINGSEARCHES = "Trending searches";
