+++ Frontend/src/components/serviceDropdown/service-dropdown.jsx
@@ -1,7 +1,10 @@
-import modalStyles from '../css/modal.module.css'
-import btnStyles from '../css/button.module.css'
-import textStyles from '../css/text.module.css'
+import modalStyles from '../../css/modal.module.css'
+import btnStyles from '../../css/button.module.css'
+import textStyles from '../../css/text.module.css'
 import classnames from 'classnames'
+import serviceDropdownStyles from './service-dropdown.module.css'
+import * as consts from './service-dropdown.consts.js'
+import trendingSVG from '../../assets/trending.svg'
 
 function ServiceDropdown(btnTextArr){
     let textArr = ['Professional Cleaning','Salon','Electricians','Plumbers','Carpenters','Washing machine repair'];
@@ -9,15 +12,13 @@ function ServiceDropdown(btnTextArr){
     let textList = [];
     let textclasses = classnames(textStyles['sub-title'],textStyles['sub-title-sm'])
     let btnclasses = classnames(btnStyles['btn'],btnStyles['small-btn-non-bold'])
-    let w = "13px";
-    let h = "13px";
+
     textArr.forEach(element => {
         textList.push(
         <>
             <button className={btnclasses}>
-                <svg height={h} width={w} xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                    <path stroke-linecap="round" stroke-linejoin="round" d="M2.25 18 9 11.25l4.306 4.306a11.95 11.95 0 0 1 5.814-5.518l2.74-1.22m0 0-5.94-2.281m5.94 2.28-2.28 5.941" />
-                </svg> {element}
+                <img src={trendingSVG} alt='trend' className={serviceDropdownStyles['searchsvg']}></img>
+                 {element}
                 </button>
         </>
         );
@@ -25,7 +26,7 @@ function ServiceDropdown(btnTextArr){
 
     return <><div className={modalStyles['modal-vertical-container']}>
         <ul>               
-            <div className={textclasses}> Trending searches </div>
+            <div className={textclasses}> {consts.TRENDINGSEARCHES} </div>
             {textList}
         </ul>
         </div>
