+++ Frontend/src/components/serviceDropdown/service-dropdown.module.css
@@ -1,0 +1,4 @@
+.searchsvg{
+    width: 13px;
+    height: 13px;
+}
