+++ Frontend/src/components/shared/typing-animation.consts.js
@@ -1,0 +1,1 @@
+export const SEARCHFOR = 'Search for ';
