+++ Frontend/src/main.jsx
@@ -2,8 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
-import ServiceDropdown from './components/service-dropdown.jsx'
-import SiteHeader from './components/header.jsx'
+import SiteHeader from './components/header/header.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
